Split main function from elf-parser.c, allows elf-parser.o to be used as a library
