+++ Makefile
@@ -1,7 +1,7 @@
 CC=gcc
 CFLAGS= -Wall -g -I.
 DEPS = elf-parser.h
-OBJ = disasm.o elf-parser.o
+OBJ = disasm.o elf-parser.o elf-parser-main.o
 
 %.o: %.c $(DEPS)
 	$(CC) -c -o $@ $< $(CFLAGS)
